requirements.txt
mkdocs-material>=9.5
mkdocs-static-i18n>=1.2
pymdown-extensions>=10.0
jsonschema>=4.17
PyYAML>=6.0
